Move/cleanup crafted/catalyst displays
