+++ src/app/item-popup/WeaponCatalystInfo.m.scss
@@ -1,19 +1,13 @@
 @use '../variables.scss' as *;
 
 .catalystProgressTitle {
-  margin-left: 4px;
-  padding-top: 2px;
-  padding-bottom: 2px;
-
-  color: $exotic;
-  font-weight: bold;
+  margin-bottom: 4px;
 }
 
 .catalystProgress {
-  display: flex;
-  flex-direction: column;
+  composes: flexColumn from '../dim-ui/common.m.scss';
   background: var(--theme-item-popup-panel-bg);
-  gap: 4px;
+  padding: 4px 10px;
 
   // Add spacer when adjacent to crafted weapon progress
   :global(.crafted-progress) + & {
@@ -23,9 +17,4 @@
   :global(.objective-progress) {
     background: var(--theme-item-popup-progress-bar-bg);
   }
-}
-
-.catalystProgressSection {
-  padding: 4px 10px 4px;
-  flex-grow: 1;
 }
