+++ src/app/item-popup/WeaponCatalystInfo.m.scss.d.ts
@@ -1,0 +1,8 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'catalystProgress': string;
+  'catalystProgressTitle': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
