+++ src/app/item-popup/WeaponCraftedInfo.m.scss
