+++ src/app/item-popup/WeaponDeepsightInfo.m.scss
