+++ src/app/item-popup/WeaponDeepsightInfo.tsx
